backend/requirements.txt
# Phase II Backend Dependencies
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database and ORM
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
bcrypt>=4.1.2

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development tools
ruff>=0.2.0
black>=24.1.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
